move dylib APIs to a better location
