+++ runtime/include/tt/runtime/detail/ttnn.h
@@ -131,11 +131,6 @@ std::vector<Tensor> runProgram(::ttnn::MeshDevice &meshDevice,
                                std::uint32_t programIndex,
                                std::vector<::ttnn::Tensor *> const &inputs);
 
-bool compareOuts(std::vector<Tensor> &lhs, std::vector<Tensor> &rhs);
-
-std::vector<Tensor> runSoProgram(void *so, std::string func_name,
-                                 std::vector<Tensor> inputs, Device device);
-
 } // namespace tt::runtime::ttnn
 
 #endif
