+++ runtime/include/tt/runtime/runtime.h
@@ -111,13 +111,6 @@ Event submit(Device deviceHandle, Binary executableHandle,
              std::uint32_t programIndex, std::vector<Tensor> const &inputs,
              std::vector<Tensor> const &outputs);
 
-void *openSo(std::string path);
-
-std::vector<Tensor> runSoProgram(void *so, std::string name,
-                                 std::vector<Tensor> inputs, Device device);
-
-bool compareOuts(std::vector<Tensor> &lhs, std::vector<Tensor> &rhs);
-
 } // namespace tt::runtime
 
 #endif
