+++ runtime/lib/runtime.cpp
@@ -489,59 +489,4 @@ Event submit(Device deviceHandle, Binary executableHandle,
   LOG_FATAL("runtime is not enabled");
 }
 
-void *openSo(std::string path) {
-#if defined(TT_RUNTIME_ENABLE_TTNN)
-  if (getCurrentRuntime() == DeviceRuntime::TTNN) {
-    // Open handle to shared object
-    //
-    void *handle = dlopen(path.c_str(), RTLD_LAZY);
-    if (!handle) {
-      LOG_FATAL("Failed to load shared object: ", dlerror());
-    }
-
-    // Clear any existing errors
-    //
-    dlerror();
-
-    return handle;
-  }
-#endif
-
-#if defined(TT_RUNTIME_ENABLE_TTMETAL)
-  if (getCurrentRuntime() == DeviceRuntime::TTMetal) {
-    LOG_FATAL("not implemented");
-  }
-#endif
-  LOG_FATAL("runtime is not enabled");
-}
-
-std::vector<Tensor> runSoProgram(void *so, std::string name,
-                                 std::vector<Tensor> inputs, Device device) {
-#if defined(TT_RUNTIME_ENABLE_TTNN)
-  return ::tt::runtime::ttnn::runSoProgram(so, name, inputs, device);
-#endif
-
-#if defined(TT_RUNTIME_ENABLE_TTMETAL)
-  if (getCurrentRuntime() == DeviceRuntime::TTMetal) {
-    LOG_FATAL("not implemented");
-  }
-#endif
-  LOG_FATAL("runtime is not enabled");
-}
-
-bool compareOuts(std::vector<Tensor> &lhs, std::vector<Tensor> &rhs) {
-#if defined(TT_RUNTIME_ENABLE_TTNN)
-  if (getCurrentRuntime() == DeviceRuntime::TTNN) {
-    return ::tt::runtime::ttnn::compareOuts(lhs, rhs);
-  }
-#endif
-
-#if defined(TT_RUNTIME_ENABLE_TTMETAL)
-  if (getCurrentRuntime() == DeviceRuntime::TTMetal) {
-    LOG_FATAL("not implemented");
-  }
-#endif
-  LOG_FATAL("runtime is not enabled");
-}
-
 } // namespace tt::runtime
